Create  How to count the number of 1s in the binary representation of numbers from 0 to n.  Write its algorithm, program. Find its time and space complexities. Explain with  suitable example.
Algorithm:

We can use a dynamic programming approach to solve this problem efficiently. Let dp[i] store the total number of set bits in the binary representations of numbers from 0 to i.

Base Case: dp[0] = 0 (binary of 0 is '0', which has zero 1s).

Iteration: Iterate from i = 1 to n. For each i:

If i is a power of 2 (i.e., i & (i - 1) == 0):

Let p be the exponent such that i = 2^p.
The numbers from 0 to i - 1 have already been considered.
The numbers from i to 2i - 1 will have the same pattern of set bits as 0 to i - 1, but with an additional '1' at the p-th bit position.
The number of integers from 0 to i - 1 is i. Each of these, when we add i to it, will have one more set bit at the p-th position.
The total number of set bits from 0 to i - 1 is dp[i - 1].
So, dp[i] = dp[i - 1] + (i / 2) where i/2 represents the number of set bits contributed by the newly set MSB for numbers from i/2 to i-1 when considering numbers from 0 to i-1.
Alternatively, when i is a power of 2, the number of set bits from 0 to i-1 is i * p / 2. When we go from 0 to i (inclusive), we add one more set bit for i itself. A more direct formula when i is a power of 2 (2 
p
 ): dp[i] = dp[i - 1] + 1. And dp[2^p - 1] = p * 2^(p-1). So, dp[2^p] = p * 2^(p-1) + 2^p.
If i is not a power of 2:

Find the largest power of 2 less than or equal to i. Let it be power_of_2.
The numbers from 0 to power_of_2 - 1 have dp[power_of_2 - 1] set bits.
The numbers from power_of_2 to i have a '1' at the most significant bit (the bit corresponding to power_of_2). There are i - power_of_2 + 1 such numbers.
The remaining bits in the numbers from power_of_2 to i follow the pattern of set bits from 0 to i - power_of_2.
So, dp[i] = dp[power_of_2 - 1] + (i - power_of_2 + 1) + dp[i - power_of_2].
Return dp[n].